basic nft implementation

+++ common/src/helpers.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Ternoa.  If not, see <http://www.gnu.org/licenses/>.
 
-pub fn check_bounds<T>(src_len: usize, min_len: (u16, T), max_len: (u16, T)) -> Result<(), T> {
+pub fn check_bounds<T>(src_len: usize, min_len: (u32, T), max_len: (u32, T)) -> Result<(), T> {
 	if src_len < min_len.0 as usize {
 		return Err(min_len.1)
 	}
